Move CACHEDIR into $HOME/.cache/EPICS
I order to be able to test the build scripts locally, before
pushing them somewhere and waiting some minutes, move them
into its own sub-directory under $HOME/.cache:
$HOME/.cache/EPICS

This allows to run `rm -rf $HOME/.cache/EPICS` to get a clean build,
and works on a remote machine as well.
If it does not exit, create it.
Side note: The definition of CACHEDIR is both in build.sh and prepare.sh,
that needs to be cleaned up in a seperate commit.