Introduce sha_accumulated to rebuild on dependend modules
The current make system does not detect, when e.g. asyn has a new commit,
and as a result, motor has to be rebuild.
Currently asyn is rebuild, but not motor.

Solution a) would be to define depencies for each module.
However, I feel that this is more error-prone than expected.
motor can be build with and without IPAC, and which module
should "own" this decision ?

Solution b) takes a different approach.
When the list of modules is build, the resulting sha sum is depending
on the module build before:

SHA_NEW=$(echo $SHA_PREV_MODS $HEAD | shasum | tr -d ' -')

So if asyn gets a new sha value on the master branch, the sha for motor
does not match because of this, and motor is rebuild.
The disadvantage is that e.g. calc can be rebuild because asyn
is new, but I think this is not a big deal.

For that reason we go for b) and introduce a variable "sha_accumulated"
which "accumulates all sha sums".